lib: add context and reporting
The context will import the asserts library and override assert(). That
will then be used to calculate the number of asserts that pass and fail
and gather up a stacktrace of each one.

At the end of the run each test runner will report on how many tests
passed and failed with a grand total given at the end.